chore: rename metric to instrument
