+++ experimental/packages/opentelemetry-sdk-metrics-base/src/Instruments.ts
@@ -14,6 +14,11 @@
  * limitations under the License.
  */
 
+import * as api from '@opentelemetry/api';
+import * as metrics from '@opentelemetry/api-metrics';
+import { Measurement } from './Measurement';
+import { Meter } from './Meter';
+
 export enum InstrumentType {
     COUNTER = 'COUNTER',
     HISTOGRAM = 'HISTOGRAM',
@@ -21,4 +26,58 @@ export enum InstrumentType {
     OBSERVABLE_COUNTER = 'OBSERVABLE_COUNTER',
     OBSERVABLE_GAUGE = 'OBSERVABLE_GAUGE',
     OBSERVABLE_UP_DOWN_COUNTER = 'OBSERVABLE_UP_DOWN_COUNTER',
+}
+
+export class Instrument {
+    constructor(private _meter: Meter, private _name: string, private _version?: string) { }
+
+    getName(): string {
+        return this._name;
+    }
+
+    /**
+     * This only exists to stop typescript complaining about unused variable and may be removed. 
+     */
+    getVersion(): string | undefined {
+        return this._version;
+    }
+
+    aggregate(measurement: Measurement) {
+        this._meter.aggregate(this, measurement);
+    }
+}
+
+export class UpDownCounter extends Instrument implements metrics.Counter {
+    add(value: number, attributes?: api.SpanAttributes, ctx?: api.Context): void {
+        if (typeof value != 'number') {
+            api.diag.warn(`invalid type value provided to counter ${this.getName()}: ${typeof value}`);
+            return;
+        }
+
+        attributes = attributes ?? {};
+        ctx = ctx ?? api.context.active();
+
+        this.aggregate({
+            value,
+            attributes,
+            context: ctx,
+        });
+    }
+}
+
+export class Counter extends UpDownCounter implements metrics.Counter {
+    override add(value: number, attributes?: api.SpanAttributes, ctx?: api.Context): void {
+        if (value < 0) {
+            api.diag.warn(`negative value provided to counter ${this.getName()}: ${value}`);
+            return;
+        }
+
+        return super.add(value, attributes, ctx);
+    }
+}
+
+export class Histogram extends Instrument implements metrics.Histogram {
+    record(_measurement: number, _labels?: metrics.Attributes, _context?: api.Context): void {
+        throw new Error('Method not implemented.');
+    }
 }
