+++ experimental/packages/opentelemetry-sdk-metrics-base/src/MeterProvider.ts
@@ -17,9 +17,9 @@
 import * as api from '@opentelemetry/api';
 import * as metrics from '@opentelemetry/api-metrics';
 import { Resource } from '@opentelemetry/resources';
+import { Instrument } from './Instruments';
 import { Measurement } from './Measurement';
 import { Meter } from './Meter';
-import { Metric } from './Metric';
 import { MetricExporter } from './MetricExporter';
 import { MetricReader } from './MetricReader';
 import { View } from './View';
@@ -148,7 +148,7 @@ export class MeterProvider {
         }
     }
 
-    public aggregate(_meter: Meter, _metric: Metric, _measurement: Measurement) {
+    public aggregate(_meter: Meter, _metric: Instrument, _measurement: Measurement) {
         // TODO actually aggregate
 
         /**
